Initial commit of Rosetta CVS snapshot from 20090508
This was likely the latest revision

On LinuxTag 2003 in Karlsruhe, I started this project. Shortly later
this was included as one of the standard input methods into the
GPE Palmtop Environment.

What is rosetta?
----------------

Rosetta is a new handwriting recognition engine for the X Window System.
It supports multi stroke recognition and recognition for whole words /
sentences. It captures characters written in a small window with a
pointing device (mouse, tablet or pen), recognizes them and passes the
relating keypress events to the actualy focused window.
You can write multi characters at once, but they may not be connected
(like script). Single characters are recognized depending on the field
you write in (lowercase, uppercase or special characters), multiple
characters are recognizes as lower case characters .

Rosetta has been developed on Linux systems, (i386 and StrongARM), but
should be quite portable to any reasonable system with X. Rosetta
still is experimental.

Many of the ideas that have gone into the development of Rosetta are
derived from xmerlin, written by Stefan Hellkvist. To His Ideas I added
the multi stroke support and some more classification functions, resulting
in Rosetta.